Move test verification code to runtime, with minor refactoring
Because it appears impossible to verify the omap of `e`/`E`/`ge`/`gE` correctly in compile-time when paired with operator `d`. Under certain condition, the delete operation of these motions need to be invoked with `d:command`, whereas in other circumstances `dv:command`. The check of such condition can only be applied in runtime.